Cleanly exposing the tools withing the tools submodule.
